requirements.txt
aiofiles==24.1.0
alkana==0.0.3
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
distro==1.9.0
filelock==3.18.0
fsspec==2025.7.0
greenlet==3.2.3
groq==0.30.0
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
keyboard==0.13.5
langchain==0.3.26
langchain-core==0.3.69
langchain-groq==0.3.6
langchain-text-splitters==0.3.8
langsmith==0.4.8
libretranslatepy==2.1.1
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
mecab-python3==1.0.10
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.97.0
openai-whisper==20250625
opencv-python==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
plac==1.4.5
platformdirs==4.3.8
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyfiglet==1.0.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.99
PyTubeStudio==0.1.0
pytz==2025.2
pyvts==0.3.3
pywin32==311
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==1.5.0
safetensors==0.5.3
setuptools==80.9.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
SQLAlchemy==2.0.41
srt==3.5.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
translate==3.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unidic==1.1.0
urllib3==2.5.0
vosk==0.3.45
wasabi==0.10.1
websockets==15.0.1
zstandard==0.23.0
